[refactor] #91 - graphql사용했던 서비스 이름 변경 및 isp 원칙에 맞도록 분리 작업 진행
