+++ src/main/java/gitbal/backend/domain/majorlanguage/TopLanguageService.java
@@ -1,0 +1,9 @@
+package gitbal.backend.domain.majorlanguage;
+
+import org.springframework.http.ResponseEntity;
+
+public interface TopLanguageService {
+
+    ResponseEntity<String> requestTopLanguageQuery(String username);
+
+}
