+++ src/main/java/gitbal/backend/domain/user/UserInfoService.java
@@ -1,0 +1,7 @@
+package gitbal.backend.domain.user;
+
+import org.springframework.http.ResponseEntity;
+
+public interface UserInfoService {
+    ResponseEntity<String> requestUserInfo(String username);
+}
